Refresh the credentials automatically
When the return value from Amazon is a 403, wipe out the p.Credentials
object so that the next request gets a new one. This isn't perfect
because you still fail the request .. but the intention here is to
automatically recover when the credentials you had expired, and possibly
new ones have been generated and pushed to the ~/.aws/credentials file.